// components/datalens-ui/src/ui/components/DashKit/plugins/Control/types.ts
export type SelectorError = {
    code: string;
    debug: string | {requestId?: string};
    details?: {
        sources?: Record<
            string,
            {
                body?: {
                    debug: Record<string, string>;
                    message: string;
                    details: Record<string, string>;
                    code: string;
                };
                data?: {
                    ignore_nonexistent_filters: boolean;
                    fuild_guid: string;
                    where: ServerFilter[];
                };
                message?: string;
                sourceType?: string;
                status?: number;
                uiUrl?: string;
                url?: string;
                code?: string;
            }
        >;
    };
};